Organize components and fix login
