+++ components/canvas/InteractiveCanvas.tsx
