+++ components/character/AddCompetenceModal.tsx
