+++ components/character/CompetencesPanel.tsx
