+++ components/character/DescriptionPanel.tsx
