+++ components/character/EquipPanel.tsx
