+++ components/character/StatsPanel.tsx
