+++ components/chat/DiceStats.tsx
