+++ components/chat/SummaryPanel.tsx
