+++ components/dice/PopupResult.tsx
