+++ components/login/DiceBackground.tsx
