+++ components/misc/SideNotes.tsx
